requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.4
python-dotenv==1.0.1
sqlite-utils==3.36
emoji==2.10.1
pandas==2.2.2
tabulate==0.9.0
xlsxwriter
flask
requests
python-telegram-bot[webhooks]
pytz
